Move demoPollingInterval to external file to prevent bundling of test files
