+++ src/config/test_constants.ts
@@ -1,0 +1,1 @@
+export const demoPollingInterval = 0.005; // 1 000 * 0.022 = 1320 ms (1.3 sec)
