+++ src/lib/flagshipSdkHelper.ts
@@ -6,7 +6,7 @@ import { BucketingApiResponse } from '../class/bucketing/types';
 import { DecisionApiCampaign, DecisionApiResponse, DecisionApiResponseData, FlagshipVisitorContext } from '../class/flagshipVisitor/types';
 import defaultConfig, { internalConfig } from '../config/default';
 import otherSdkConfig from '../config/otherSdk';
-import { demoPollingInterval } from '../config/test';
+import { demoPollingInterval } from '../config/test_constants';
 import { FlagshipSdkConfig, IFsPanicMode, PostFlagshipApiCallback } from '../types';
 
 const checkRequiredSettingsForApiV2 = (config: FlagshipSdkConfig, log: FsLogger): void => {
